Fix comments
Fix Email preview logo
Fix Comments not being displayed
