+++ met-api/src/met_api/models/submission.py
@@ -157,41 +157,6 @@ class Submission(BaseModel):  # pylint: disable=too-few-public-methods
         return query.first()
 
     @classmethod
-    def get_by_survey_id_paginated(
-        cls,
-        survey_id,
-        pagination_options: PaginationOptions,
-        search_text='',
-        advanced_search_filters=None
-    ):
-        """Get submissions by survey id paginated."""
-        null_value = None
-        query = db.session.query(Submission)\
-            .filter(and_(Submission.survey_id == survey_id,
-                         or_(Submission.reviewed_by != 'System', Submission.reviewed_by == null_value)))\
-
-        if search_text:
-            # Remove all non-digit characters from search text
-            query = query.filter(cast(Submission.id, TEXT).like('%' + search_text + '%'))
-
-        if advanced_search_filters:
-            query = cls._filter_by_advanced_filters(query, advanced_search_filters)
-
-        sort = asc(text(pagination_options.sort_key)) if pagination_options.sort_order == 'asc'\
-            else desc(text(pagination_options.sort_key))
-
-        query = query.order_by(sort)
-
-        no_pagination_options = not pagination_options.page or not pagination_options.size
-        if no_pagination_options:
-            items = query.all()
-            return items, len(items)
-
-        page = query.paginate(page=pagination_options.page, per_page=pagination_options.size)
-
-        return page.items, page.total
-
-    @classmethod
     def get_engaged_participants(cls, engagement_id) -> List[Participant]:
         """Get users that have submissions for the specified engagement id."""
         users = db.session.query(Participant)\
