+++ met-api/src/met_api/resources/comment.py
@@ -19,6 +19,7 @@ from flask import Response, request
 from flask_cors import cross_origin
 from flask_restx import Namespace, Resource
 
+from met_api.auth import auth
 from met_api.auth import jwt as _jwt
 from met_api.models.pagination_options import PaginationOptions
 from met_api.services.comment_service import CommentService
@@ -60,7 +61,7 @@ class SurveyComments(Resource):
 
     @staticmethod
     @cross_origin(origins=allowedorigins())
-    @_jwt.requires_auth
+    @auth.optional
     def get(survey_id):
         """Get comments page."""
         try:
