+++ met-cron/requirements.txt
@@ -11,7 +11,7 @@ Mako==1.2.4
 MarkupSafe==2.1.3
 SQLAlchemy-Utils==0.41.1
 SQLAlchemy==1.3.24
-Werkzeug==2.3.4
+Werkzeug==2.3.6
 alembic==1.11.1
 aniso8601==9.0.1
 attrs==23.1.0
@@ -27,7 +27,6 @@ flask-marshmallow==0.11.0
 flask-restx==1.1.0
 gunicorn==20.1.0
 idna==3.4
-importlib-metadata==6.6.0
 itsdangerous==2.0.1
 jaeger-client==4.8.0
 jsonschema==4.17.3
@@ -50,5 +49,4 @@ thrift==0.16.0
 tornado==6.3.2
 typing_extensions==4.6.3
 urllib3==2.0.3
-zipp==3.15.0
 -e git+https://github.com/bcgov/met-public.git#egg=met-api&subdirectory=met-api
