+++ met-web/src/components/comments/admin/review/emailPreview/EmailPreview.tsx
@@ -1,7 +1,7 @@
 import { Box, Grid, Stack } from '@mui/material';
 import * as React from 'react';
 import { MetHeader4, MetBody } from 'components/common';
-import { ReactComponent as BCLogo } from 'assets/images/BritishColumbiaLogoLight.svg';
+import { ReactComponent as BCLogo } from 'assets/images/BritishColumbiaLogoDark.svg';
 import { Survey } from 'models/survey';
 import { formatDate } from 'components/common/dateHelper';
